Add build targets to server edition
Define for the server edition the same build targets as for the empty
edition, but using the correct template so that the "offline" version
(target "index") works correctly when accessed via HTTP. With this,
`tiddlywiki --build index` is equivalent to the save button.

While the process of setting up TiddlyWiki on Node.js is well documented
and easy enough, the options for  publishing such a wiki to an offline
version (scriptably, e.g. for push-to-deploy setups) are decidedly
non-obvious. With the added build steps, the user only needs to know
three simple commands:

tiddlywiki --init server
tiddlywiki --server
tiddlywiki --build index

and optionally

tiddlywiki --build static