Introduce keyboard shortcuts and refactor buttons
Quite a few changes that rippled out from one another:

1. Extend the keyboard shortcut utilities to allow named shortcuts (eg
instead of “ctrl-B” one can say “((bold))”, and the actual keyboard
shortcut is retrieved from `$:/config/KeyboardShortcuts/bold`)
2. Extend the keyboard shortcut utilities to work with lists of
shortcuts
3. Refactor the editor toolbar buttons to make them more declarative.
Now the framework takes care of the button and the dropdown, and the
button tiddler just needs to specify the content of the button, and the
title of an optional dropdown. This turns out to be a useful space
saving measure that we may want to investigate for the other toolbars
4. Add a “preview-type” dropdown for choosing the type of the preview
pane (this button is only shown if there are multiple preview panes
installed)

Apologies to those who have already created toolbar buttons, and will
need to change them in line with this commit.