Collision between the core and the pre-release language plugin
In the language plugin there is a tiddler also called "GettingStarted" which overrides the tiddler in the core. I know the folder es-ES within the wikiFolder is provisory, to test the language plugin translation. I have checked the translation in the .tid and renamed the tiddler to momentarily avoid the collision.