Translation goes on + minor fixes
Continued translation of tiddlers and a couple minor fixes:

The original text does not follow TW's own style policies. Translating trying to adapt to it gets a bit tricky sometimes. It will require an in-depth proofreading afterwards, as well as a revision of TW documentation, in order to adapt to the policies.

-I chopped Saving on TiddlySpot: Moved the first paragraph to the missing tiddler TiddlySpot and then transcluded its content back.