Provisorily copied languages/es-ES into editions/es-ES/tiddlers
In order to force the load of the new translated strings and test the es-ES plugin translation in situ, I've provisorily copied the plugin folder into the tiddlerStore. This folder will be deleted when the port to the spanish edition will be done. 
file GettingStarted.tid of the language plugin has been renamed as GettingStarted_clash.tid, as it overrides the core GettingStarted.tid