Update TagTemplate.tid
My request is to add the Open All feature without drag and drop until the link widget can be extended so in the future it can be dragged as a group of tiddlers but for now at least get the feature of opening all with a particular tag available since it is so desired by many. This may be a duplicate of 2513 or even 192. Please forgive me since I may not know enough to be qualified to try to submit pull requests.