Add this.document to ES2016 documentation (#2314)
In the sample example for using ES2016 it used the
`$tw.utils.domMaker()`. I just found out that if this code is executed
by Node.js instead of in the browser that it doesn't work because
`window.document` doesn't exist. The expectation is that widgets pass in
the fake `this.document`.

I updated the example to reflect this.