Add support to read tiddler specs from JS
If the system finds `tiddlywiki.files.js` it will execute it and take
the return result as the `fileInfo`. Otherwise it will parse the JSON
like normal.

It is possible for a node module to return an Object instead of a
function; however, since the few examples in the wild where I've seen
this (i.e. Ember.js and Broccoli) they use a function. I've found that a
simple function offered the most flexibility at the expense of a tiny
bit more boilerplate.