Allow savers to receive options and variables
This pull request allows sending options to the savers. This may be useful if you want to configure the way the saver behaves when you send the save event, or even to allow communication between savers (or the same saver) if you save mechanism happens in two phases/steps. 

The previous code was a bit confusing because it was saving `optioins.variables`  into variables just for rebuilt the same object later `{variables: { filename: variables.filename}}`.