Update render.tid
This is just a small proposal to enhance the user experience while reading the `--render` documentation. It's my way of saying it could be better by providing an example. This adds a link that opens the advanced search with a filter to list all the template tiddlers.
I made this from memory, I know it works but probably can be improved.