Add name field to server start hook
This allows other server implementations (such as TiddlyServer) to provide custom support for features a plugin wishes to add by access the Node server object. The main example is WebSockets, since many data folders are loaded into TiddlyServer and therefore cannot extend the NodeJS server object. 

Even if they could, TiddlyServer uses two separate server instances. So it's better to do it like this. TiddlyServer uses the SimpleServer router so the first argument of the hook still works fine and is not affected.