Comment out code that removes deleted tiddlers
I don't think this is the right way to go about it because there could be a discrepancy between the server and the client. I think a better way would be to use server-sent events to actually send a delete message to the browser and let the browser decide whether to delete it or not.