Contribute webmodule code from TiddlyServer
This is the code that I use to load data folders. I wrote it myself, and have now modified it slightly for using TiddlyWiki directly. It is useful for loading a data folder as a route in express or other route-based servers but must have access to the original request and response object. 

Requests are handled by calling `server.requestHandler(req, res)`.