[doc] Update ListWidget.tid
Move section "Content and Attributes" to before Examples to be consistent with several other widget docs (e.g RevealWidget). 

The Examples section is thus moved lower which is good because it would allow us to add more examples. I intend to add example of how to make html tables using the ListWidget.