Use as a real filter
I misunderstood the behaviour: I thought that I need to collect tiddlers
(initialize a list) and return with them to let the `+[...]` filter
expression collect the elements common with the previous expression, but
instead of this I have to filter the input tiddlers by checking that
it's a member of the family or not.

The new syntax is:

  [kindred:<field>[<tiddler_from_family>]]
  [kindredup:<field>[<tiddler_from_family>]]
  [kindreddown:<field>[<tiddler_from_family>]]

Because of this filter behaviour, the "base tiddler" (passed as operand)
has to be included in to the output list (I excluded it earlier).

WARNING: Commit introducing TODOs