Solve the problem with cyclical connections
If there is a cyclical family tree (the "bottom" (C) element is tagging
the "top" (A)) and a branch is derived from one of the members (B), then
the branch (B2) is not always found in the filter.

  A
  |\
  | B
  |/ \
  C   B2

During the crawling of the tree, it first adds all the elements to the
list, but does not scan them in the other direction: if it finds a that
has been checked already, it will be skipped to prevent endless loops.

To find all the elements, but avoid endless loops, I'll pick
up the results in two separate lists and then combine them.

WARNING: Commit introducing TODOs