tree macro - fix for double quotes
This is trying to fix the problem with double quotes (") in branches/leaves names.
One can try to replicate the problem by creating a system tiddler like this: $:/_temp/"abc" or $:/"_temp"/abc and then see how it appears in the "Explorer" sidebar tab.
I'm using double quotes in standard tiddlers titles in my day to day work and this problem appears when I create state tiddlers automatically in lists (e.g. for popups).