Documentation of Using Excise text (#2533)
* Documentation of Using Excise text 

Documentation on the Excise text function. 
I propose that the  Excise text drop down ( the tiddler: $:/language/Buttons/Excise/Caption/Tag) links to this doc. I don't have permission to (or don't know how to) edit that tiddler in github. 

I propose the following text for the drop-down tiddler:

Tag new tiddler with the title of this tiddler (If the title is changed you have to [[save and re-edit|Using Excise]] to use the new title as tag).


see the issue:
https://github.com/Jermolene/TiddlyWiki5/issues/2531

* Update Using Excise.tid

changed headings in response to discussion

* Doc for  editor toolbar buttons (wikitext)

Documentation for buttons special to the editor toolbar (wikitext)

* Update Using Excise.tid

placed it under the new [[Editor toolbar]]

* Update Using Stamp.tid

placed it under the new [[Editor toolbar]] tiddler

* Update WikiText.tid

added reference to [[Editor toolbar]]

* Update Formatting text in TiddlyWiki.tid

added reference to the [[Editor toolbar]]

* Update WikiText.tid

* Update Editor toolbar.tid

* Update Editor toolbar.tid

* Update Editor toolbar.tid

* Update Editor toolbar.tid

typo