Fix diff-widget undefined attributes bug
Calling diff_main method with undefined parameters throws a (uncaught) "Null input" error.
This patch adds default values to the widget "source" and "dst" attributes so the diff method don't throw an error when widget attributes are not defined or their value is not an string object (undefined).

The default value string has been selected to help identifying the error when there is a syntax error in the widget argument names or their value is undefined. You might want to change it as you see fit.