fix tags in viewtemplate not animating
this PR fixes the tags in the tags-viewtemplate which should animate because they're listed using storyview="pop"

the problem is that the tiddler widget refreshes if the state-hash changes, and that hash includes the tag-list

the solution used here is using two different hashes, one without the tags and another one including the tags. it makes a full refresh if the hash without the tags has changed. if only the one including the tags has changed, it waits for the animation to finish, then sets the `tiddlerTagClasses` variable to the new value and refreshes the changedTiddler's children