add "isHidden" attribute to domNode
this adds an additional `isHidden` attribute to domNodes but not like the "hidden" attribute after the animation is finished, but right after the state is triggered

this isn't used in the core, but it's useful if listening for the change of this hidden attribute. in that case it's better to be informed instantly when it's about to get the hidden attribute. so one listens for the `isHidden` attribute change

a proposal, but a very useful one for me :smile: