Update wikitext-macros.tid
Added new macro wikitext-example-compact to support the publishing of Widget variants in a more compact form.

Initially to support changes to the ViewWidget Documentation, this compact form may be used in other Widget documentation.

Variants demonstrate alternative forms of a specific widget or macro with the ability for users to copy to clipboard and use in their own wiki. In some widgets that have multi-functions there is no single form of the widget, and demonstrating some variations highlights this.