Update Filter Operators.tid
To add an additional paragraph at the top to direct the user to common uses of Filters, in ListWidget, SetWidget and Filtered Transclusions.

Please add to this if there is common uses elsewhere for filter. 

By providing this guidance to users at the early stage of learning they will know the key places in which Widgets use filters thus have the opportunity to experiment and learn more about filters, Otherwise they remain an abstract concept.