Cast count to a number for comparison
Prior to this change, the `count` provided was a string. However, our use for it is as a number. Though there are some cases where type coercion is desired in this case it conflates the readbility of the code.

This change casts the `count` to a number before using it in numeric comparisons.

Fixes issue #3912