Solve the p problem by removing 19 characters from a regex :-)
Proof of concept
----------------

After applying this change, copypaste the html source of a Wiki from the Developer tools into a text editor and search for `<p><div`. There will be no search results. Before, with eg. the $:/ControlPanel tiddler open, there will be six results.

What it does
------------

Optionally install the Plugin 'Tools for exploring the internals of TiddlyWiki', in order to see the generated html in the preview pane (choose 'raw HTML' from the drop down).

Then put this into a tiddler:

\whitespace trim

<y>...</y>

<y>
...
</y>

<y>

...

</y>

The resulting HTML before and after the change:

<p><y>...</y></p><p><y>...</y></p><y><p>...</p></y> (before)
<p><y>...</y></p><y>...</y><y><p>...</p></y> (after)

After the change the second y element will not get wrapped in a p, still its childs will be parsed in inline mode. The other two block types get parsed like before.

We can now have every syntax, <p><y>...</y></p> (inline mode), as well as <y>...</y> (parent block, child inline mode), as well as <y><p>...</p></y> (block mode). Before we could just choose between <p><y>...</y></p> (inline mode) and <y><p>...</p></y> (block mode).

It is an extension to the current procedere, which does not wrap the element in a p when there are two newlines after the opening tag. But these two newlines will also make the children get wrapped in a p, which is the reason for the current problem.

By adding the ruling that one newline after the opening tag will also not wrap the element in a p, but the childs will still get parsed in inline mode, we solve the problem.

Block parsing of the childs of a child can still be achieved by adding the usual two newlines after the opening tag of the child. For example ...

\whitespace trim

<a>
<b>

c

</b>
</a>

... becomes ...

<a><b><p>c</p></b></a>

A change of the existing template tiddlers is not necessary.