Clarification: Rename variable and extend doc
Explain the difference between the two code locations where we check for immediately following linebreaks.

In the first location we check if the CHILDREN shall be parsed in block or inline mode. Two (or more) linebreaks will result in block mode.

In the second we check if the PARENT shall be parsed in block or inline mode. One (or more) linebreaks will result in block mode.