Self closing elements and elements without closing tag, followed by EOS
Remove '|$' ('or end of input') from the regexes.

This will parse a self closing element (and its children) as inline run, when it is immediately followed by the end of the string. If there is a linebreak it (and its children) will be parsed as block run.

Same true for elements without a closing tag, followed by the end of string.

This is identical to the current behaviour of TiddlyWiki.

'<y/>' --> '<p><y></y></p>'
'<y/>\n' --> '<y></y>'

'<y>' --> '<p><y></y></p>'
'<y>\n' --> '<y></y>'

This results in 'node tiddlywiki ./editions/test' producing 3 instead of 9 errors.