Skip explicit linebreaks, for nicer formatting
Because of the new ruling introduced in the 'Change regex ...' commit, we are not allowed to have a newline after the opening tag, when we want the element to be wrapped in a p. For example

<button>text</button>

will wrap the button in a p, but

<button>
text
<button>

will not.

This leads to the problem that we have to use ugly formatting in situations where we have chained elements. For example, in the following situation, when we want the outermost i element to be wrapped in a p, we need to write

<i><want>
<to>
be wrapped in a p
</to>
</want>
</i>

and we can not write the intuitive

<i>
<want>
<to>
be wrapped in a p
</to>
</want>
</i>

To resolve this we introduce that <backslash><newline> directly after an opening tag will be removed (and only then). So we can now write

<i>\
<want>
<to>
be wrapped in a p
</to>
</want>
</i>

and this is exactly equivalent to

<i><want>
<to>
be wrapped in a p
</to>
</want>
</i>